ner-service/pyproject.toml
[tool.poetry]
name = "text_matching_api"
version = "0.1.0"
description = "match natural language"
authors = ["Liusha-He <liusha.8852@yahoo.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
fastapi = "<=0.74.0"
pydantic = "^1.9.0"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
pyspellchecker = "^0.6.3"
requests = "^2.27.1"
nltk = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
flake8 = "^4.0.1"
black = "^22.1.0"
coverage = "^6.3.2"
Pattern = "^3.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
